Make Announcement Feature

+++ src/main/java/com/example/greenpass/v1/Announcement/repositories/AnnouncementRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.greenpass.v1.Announcement.entities.Announcement;
 
-public interface AnnouncementRepository extends JpaRepository<Announcement, Long> {
+public interface AnnouncementRepository extends JpaRepository<Announcement, Integer> {
 
 }
